+++ eggs/Naaya/eggs/Naaya/Products/Naaya/adapters.py
@@ -147,11 +147,12 @@ class FolderMetaTypes(object):
         Uses the `use default` logic, always returns list-type
 
         """
-        if getattr(self.context, 'folder_meta_types', None) is None:
-            # None or missing means use global settings
+        folder_meta_types = getattr(self.context, 'folder_meta_types', ())
+        if folder_meta_types is None:
+            # None means use global settings
             return list(self.context.getSite().adt_meta_types)
         else:
-            return list(self.context.folder_meta_types)
+            return list(folder_meta_types)
 
     def set_values(self, meta_types):
         """